Доработана логика создания и редактирования трекера; Сделан градиент на экране статистики; Доработана логика отображения выполненных трекеров.
