+++ Tracker.xcworkspace/contents.xcworkspacedata
@@ -1,6 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Workspace
    version = "1.0">
+   <Group
+      location = "container:"
+      name = "Statistics">
+   </Group>
    <FileRef
       location = "group:Tracker.xcodeproj">
    </FileRef>
