+++ Tracker/CoreData/TrackerRecordStore.swift
@@ -14,12 +14,16 @@ enum TrackerRecordError: Error {
     case decodingErrorInvalidRecord
 }
 
+protocol TrackerRecordDelegate: AnyObject {
+    func updateCountCompletedTracker()
+}
+
 final class TrackerRecordStore: NSObject {
     
     private let colorAndDayMarshalling = ColorAndDayMarshalling.shared
     private let context: NSManagedObjectContext
     private var fetchedResultsController: NSFetchedResultsController<TrackerRecordCoreData>!
-    static let DidChangeNotification = Notification.Name("TrackerRecord")
+    weak var delegate: TrackerRecordDelegate?
     
     static let shared = TrackerRecordStore()
     
@@ -46,7 +50,7 @@ final class TrackerRecordStore: NSObject {
         record.date = date
         try context.save()
         try updateResult()
-        NotificationCenter.default.post(name: TrackerRecordStore.DidChangeNotification, object: self)
+        delegate?.updateCountCompletedTracker()
     }
     
     func getRecord() throws -> [TrackerRecord] {
@@ -80,7 +84,7 @@ final class TrackerRecordStore: NSObject {
         context.delete(record)
         try? context.save()
         try updateResult()
-        NotificationCenter.default.post(name: TrackerRecordStore.DidChangeNotification, object: self)
+        delegate?.updateCountCompletedTracker()
     }
     
     func updateResult() throws {
